♻️  Move context tools to domain
customer data warning seems pretty domain specific already and we will need to add an event listener for the future stored context manager
