+++ packages/core/src/domain/context/contextManager.spec.ts
@@ -1,9 +1,9 @@
 import type { Clock } from '../../../test'
 import { mockClock } from '../../../test'
-import { display } from '../display'
+import { display } from '../../tools/display'
 import { BYTES_COMPUTATION_THROTTLING_DELAY, createContextManager } from './contextManager'
 import { CUSTOMER_DATA_BYTES_LIMIT } from './heavyCustomerDataWarning'
-import { CustomerDataType } from './serializationConstants'
+import { CustomerDataType } from './contextConstants'
 
 describe('createContextManager', () => {
   let clock: Clock
