+++ packages/core/src/domain/context/heavyCustomerDataWarning.spec.ts
@@ -1,6 +1,6 @@
-import { display } from '../display'
+import { display } from '../../tools/display'
 import { CUSTOMER_DATA_BYTES_LIMIT, warnIfCustomerDataLimitReached } from './heavyCustomerDataWarning'
-import { CustomerDataType } from './serializationConstants'
+import { CustomerDataType } from './contextConstants'
 
 describe('warnIfCustomerDataLimitReached', () => {
   let displaySpy: jasmine.Spy<typeof display.warn>
