+++ packages/core/src/domain/context/heavyCustomerDataWarning.ts
@@ -1,6 +1,6 @@
-import { ONE_KIBI_BYTE } from '../utils/byteUtils'
-import { display } from '../display'
-import { CustomerDataType } from './serializationConstants'
+import { ONE_KIBI_BYTE } from '../../tools/utils/byteUtils'
+import { display } from '../../tools/display'
+import { CustomerDataType } from './contextConstants'
 
 // RUM and logs batch bytes limit is 16KB
 // ensure that we leave room for other event attributes and maintain a decent amount of event per batch
